Create LLM service - Deprecate AI service  (#868)
Add LLM service for ease of use of LLMs inside of gp-api

The previous ai.service is rather convoluted with mixed in logic of prompts and opinions. It would require a large blast radius refactor. I put a deprecation notice on ai.service and create llm.service to be used.

llm.service is a separation of concerns and only supports necessary functionality for chat completion, model fallbacks, token caching, and tool usages.